Routines to get reforigin data etc.

+++ scripts/pdbEd.py
@@ -4,8 +4,15 @@ Standalone scripts to manipulate PDB files
 '''
 
 # python imports
+import argparse
 import os
+import sys
 import tempfile
+
+
+# Get the path to the ample python directory and add it to our path
+apdir = os.path.abspath( os.path.join( os.path.dirname(__file__),"../python" ) )
+sys.path.append( apdir )
 
 # our imports
 import ample_util
@@ -16,16 +23,89 @@ import pdb_edit
 #logging.basicConfig()
 #logging.getLogger().setLevel(logging.DEBUG)
 
+def extract_chain( inpdb, outpdb, chainID=None, newChainID=None ):
+    """Extract chainID from inpdb and renumner"""
+    
+    if not newChainID:
+        newChainID = chainID
+    
+    logfile = outpdb+".log"
+    cmd="pdbcur xyzin {0} xyzout {1}".format( inpdb, outpdb ).split()
+    stdin="""lvchain {0}
+renchain {0} {1}
+sernum
+""".format( chainID, newChainID )
+    retcode = ample_util.run_command(cmd=cmd, logfile=logfile, directory=os.getcwd(), dolog=False, stdin=stdin)
+    
+    if retcode == 0:
+        # remove temporary files
+        os.unlink(logfile)
+        
+    return
 
+def get_info( inpdb ):
+    PE = pdb_edit.PDBEdit()
+    info = PE.get_info( inpdb )
+    return info
+
+def keep_matching( refpdb, targetpdb, outpdb ):
+    
+    tmp1 = tmpFileName()
+    tmp1+=".pdb" # pdbcur insists names have a .pdb suffix  
+    
+    PE = pdb_edit.PDBEdit()
+    PE.keep_matching( refpdb, targetpdb, tmp1 )
+    
+    # now renumber with pdbcur
+    logfile = tmp1+".log"
+    cmd="pdbcur xyzin {0} xyzout {1}".format( tmp1, outpdb ).split()
+    stdin="""sernum
+"""
+    retcode = ample_util.run_command(cmd=cmd, logfile=logfile, directory=os.getcwd(), dolog=False, stdin=stdin)
+    
+    if retcode == 0:
+        # remove temporary files
+        os.unlink(tmp1)
+        os.unlink(logfile)
+    
+    return retcode
+
+def reforigin_rmsd( refpdb, inpdb, outpdb, DMAX=100 ):
+    
+    logfile = outpdb+".log"
+    cmd="reforigin xyzin {0} xyzref {1} xyzout {2} DMAX {3}".format( refpdb, inpdb, outpdb, DMAX ).split()
+    
+    retcode = ample_util.run_command(cmd=cmd, logfile=logfile, directory=os.getcwd(), dolog=False)
+    
+    if retcode != 0:
+        raise RuntimeError, "Error running command: {0}".format( " ".join(cmd) )
+    
+    # Parse logfile to get RMSD
+    rms = None
+    for line in open( logfile, 'r' ):
+        if line.startswith("RMS deviation:"):
+            rms = float( line.split()[-1] )
+            break
+    
+    if not rms:
+        raise RuntimeError, "Error extracting RMS from logfile: {0}".format( logfile )
+    
+    return rms
+
+def tmpFileName():
+    """Return a filename for a temporary file"""
+
+    # Get temporary filenames
+    t = tempfile.NamedTemporaryFile(dir=os.getcwd(), delete=True)
+    tmp1 = t.name
+    t.close()
+    return tmp1
+    
 def to_1_std_chain( inpdb, outpdb):
     """Clean down to 1 model/chain with standard amino acids"""
     
-    
-    # Get temporary filenames
-    t = tempfile.NamedTemporaryFile(dir=os.getcwd(), delete=False)
-    tmp1 = t.name
-    t.close()
-    os.unlink(tmp1)
+
+    tmp1 = tmpFileName()
     tmp1+=".pdb" # pdbcur insists names have a .pdb suffix
     
     # Strip to one chain and standardise AA names
@@ -33,28 +113,67 @@ def to_1_std_chain( inpdb, outpdb):
     PE.to_1_std_chain(inpdb, tmp1)
     
     # Now clean up with pdbcur
-    cmd="/opt/ccp4-6.3.0/bin/pdbcur xyzin {0} xyzout {1}".format( tmp1, outpdb ).split()
+    logfile = tmp1+".log"
+    cmd="pdbcur xyzin {0} xyzout {1}".format( tmp1, outpdb ).split()
     stdin="""delsolvent
 mostprob
 """
-    retcode = ample_util.run_command(cmd=cmd, logfile=None, directory=os.getcwd(), dolog=False, stdin=stdin)
+    retcode = ample_util.run_command(cmd=cmd, logfile=logfile, directory=os.getcwd(), dolog=False, stdin=stdin)
     
-    # remove temporary file
-    os.unlink(tmp1)
+    if retcode == 0:
+        # remove temporary files
+        os.unlink(tmp1)
+        os.unlink(logfile)
     
     return retcode
 
-# root="/home/jmht/Documents/test/3PCV"
-# os.chdir(root)
-# to1stdChain(root+"/3PCV.pdb",root+"/test/3PCV_clean.pdb")
+#
+# Command-line handling
+#
+parser = argparse.ArgumentParser(description='Manipulate PDB files', prefix_chars="-")
 
-def keep_matching( refpdb, targetpdb, outpdb ):
-    PE = pdb_edit.PDBEdit()
-    PE.keep_matching( refpdb, targetpdb, outpdb )
+group = parser.add_mutually_exclusive_group()
+group.add_argument('-one_std_chain', action='store_true',
+                   help='Take pdb to one model/chain that contains only standard amino acids')
+
+group.add_argument('-keep_matching', action='store_true',
+                   help='keep matching atoms')
+
+parser.add_argument('-ref_file', type=str,
+                   help='The reference file')
+
+parser.add_argument('input_file',
+                   help='The input file - will not be altered')
+
+parser.add_argument('output_file',
+                   help='The output file - will be created')
+
+#refpdb="/home/jmht/Documents/test/3PCV/test/refmac_phaser_loc0_ALL_poly_ala_trunc_2.822761_rad_1_UNMOD_chain1.pdb"
+#targetpdb="/home/jmht/Documents/test/3PCV/test/3PCV_clean.pdb"
+#outpdb="/home/jmht/Documents/test/3PCV/test/matching1.pdb"
+
+# refpdb="/home/jmht/Documents/test/3U2F/molrep/refine/refmac_molrep_loc0_ALL_poly_ala_trunc_0.21093_rad_2_UNMOD.pdb"
+# targetpdb="/home/jmht/Documents/test/3U2F/test/3U2F_clean.pdb"
+# outpdb="/home/jmht/Documents/test/3U2F/test/matching.pdb"
+# keep_matching( refpdb, targetpdb, outpdb )
+
+#to_1_std_chain("/home/jmht/Documents/test/3U2F/3U2F.pdb","/home/jmht/Documents/test/3U2F/test/3U2F_clean.pdb")
+
+
+if "__name__" == "__main__":
+    args = parser.parse_args()
     
-    # now renumber
-
-refpdb="/home/jmht/Documents/test/3PCV/test/refmac_phaser_loc0_ALL_poly_ala_trunc_2.822761_rad_1_UNMOD_chain2.pdb"
-targetpdb="/home/jmht/Documents/test/3PCV/test/3PCV_clean.pdb"
-outpdb="/home/jmht/Documents/test/3PCV/test/matching2.pdb"
-keep_matching( refpdb, targetpdb, outpdb )
+    # Get full paths to all files
+    args.input_file = os.path.abspath( args.input_file )
+    if not os.path.isfile(args.input_file):
+        raise RuntimeError, "Cannot find input file: {0}".format( args.input_file )
+    args.output_file = os.path.abspath( args.output_file )
+    if args.ref_file:
+        args.ref_file = os.path.abspath( args.ref_file )
+        if not os.path.isfile(args.ref_file):
+            raise RuntimeError, "Cannot find ref file: {0}".format( args.ref_file )
+    
+    if args.one_std_chain:
+        to_1_std_chain( args.input_file, args.output_file )
+    elif args.keep_matching:
+        keep_matching( args.ref_file, args.input_file, args.output_file )
